Guard against missing fields in the stats file, or throw
Also, make sure that `props.onLoaded` is always called before the `setState` inside FileInputContainer. This way we know that the file was validated before saving our history.

Lots of similar catch-blocks in here. Will need to make the UI better for people (`alert()` sucks now) but we can iterate on that depending on how frequent this is.