Ignore integration tests during normal runs.
There's an issue where using *.skip when a test has snapshots causes the snapshots to be marked for removal. Makes it hard to use (abuse) skip in the way I was doing.
So instead I've just ignored the path to integration tests, this'll be a weird way to have two suites going forward.